backend/requirements.txt
# =============================================================================
# EV Life Manager - Backend Dependencies
# Python 3.12+
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# -----------------------------------------------------------------------------
# Data Validation & Serialization
# -----------------------------------------------------------------------------
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# -----------------------------------------------------------------------------
# Environment & Configuration
# -----------------------------------------------------------------------------
python-dotenv==1.0.0

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx==0.26.0
aiohttp==3.9.1

# -----------------------------------------------------------------------------
# Background Tasks & Queue
# -----------------------------------------------------------------------------
celery==5.3.4
redis==5.0.1

# -----------------------------------------------------------------------------
# AI/ML Libraries
# -----------------------------------------------------------------------------
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
tensorflow>=2.16.1
joblib==1.3.2

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2023.3

# -----------------------------------------------------------------------------
# Monitoring & Logging
# -----------------------------------------------------------------------------
sentry-sdk[fastapi]==1.39.2

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# -----------------------------------------------------------------------------
# AWS Integration
# -----------------------------------------------------------------------------
boto3==1.34.20
botocore==1.34.20

# -----------------------------------------------------------------------------
# External API Clients
# -----------------------------------------------------------------------------
googlemaps==4.10.0
requests==2.31.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-slugify==8.0.1
pillow==10.2.0
